backend/requirements.txt
# Backend
fastapi==0.115.*
uvicorn[standard]==0.34.*
sqlalchemy[asyncio]==2.0.*
alembic==1.14.*
psycopg2-binary==2.9.*
pydantic==2.10.*
pydantic-settings==2.7.*
python-dotenv==1.0.*
python-multipart==0.0.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*

# RAG / AI (install now so imports don't break; used from Week 3+)
qdrant-client==1.13.*
sentence-transformers==3.4.*
langchain==0.3.*
langchain-community==0.3.*
langchain-text-splitters==0.3.*
langchain-ollama==0.3.*
pymupdf==1.25.*

# Reporting (Week 8)
reportlab==4.2.*
matplotlib==3.10.*

# Testing
pytest==8.3.*
httpx==0.28.*
pytest-asyncio==0.25.*
